Fish-information-system/requirements.txt
argcomplete==3.3.0
asgiref==3.8.1
asn1crypto==1.5.1
bashlex==0.18
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
compiledb==0.10.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
Cython==3.0.10
distro==1.9.0
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
djangorestframework==3.15.1
drf-yasg==1.21.7
html5lib==1.1
httplib2==0.22.0
idna==3.7
importlib_metadata==7.1.0
inflection==0.5.1
itypes==1.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jeepney==0.8.0
Jinja2==3.1.4
keyring==25.2.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
mechanize==0.4.10
meson==1.4.1
more-itertools==10.3.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pip-review==1.3.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
SecretStorage==3.3.3
setuptools==70.0.0
shutilwhich==1.1.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wadllib==1.3.6
webencodings==0.5.1
wheel==0.43.0
zipp==3.19.2
